MixQ/src/MixQ/requirements.txt
flask>=3.0
flask_limiter
pywebview>=5.0
requests

